Restructure MatterNode base class and move around some methods
