+++ packages/matter.js/src/CommissioningServer.ts
@@ -9,7 +9,8 @@ import { MatterNode } from "./MatterNode.js";
 import { Ble } from "./ble/Ble.js";
 import { AttestationCertificateManager } from "./certificate/AttestationCertificateManager.js";
 import { CertificationDeclarationManager } from "./certificate/CertificationDeclarationManager.js";
-import { Attributes, Events } from "./cluster/Cluster.js";
+import { Attributes, Cluster, Commands, Events } from "./cluster/Cluster.js";
+import { ClusterClientObj } from "./cluster/client/ClusterClientTypes.js";
 import { AccessControlCluster } from "./cluster/definitions/AccessControlCluster.js";
 import {
     AdministratorCommissioning,
@@ -47,7 +48,7 @@ import { EndpointNumber } from "./datatype/EndpointNumber.js";
 import { FabricIndex } from "./datatype/FabricIndex.js";
 import { VendorId } from "./datatype/VendorId.js";
 import { Aggregator } from "./device/Aggregator.js";
-import { Device } from "./device/Device.js";
+import { Device, RootEndpoint } from "./device/Device.js";
 import { Endpoint } from "./device/Endpoint.js";
 import { Logger } from "./log/Logger.js";
 import { MdnsBroadcaster } from "./mdns/MdnsBroadcaster.js";
@@ -55,7 +56,7 @@ import { MdnsInstanceBroadcaster } from "./mdns/MdnsInstanceBroadcaster.js";
 import { MdnsScanner } from "./mdns/MdnsScanner.js";
 import { UdpInterface } from "./net/UdpInterface.js";
 import { InteractionServer } from "./protocol/interaction/InteractionServer.js";
-import { TypeFromBitSchema, TypeFromPartialBitSchema } from "./schema/BitmapSchema.js";
+import { BitSchema, TypeFromBitSchema, TypeFromPartialBitSchema } from "./schema/BitmapSchema.js";
 import {
     CommissionningFlowType,
     DiscoveryCapabilitiesBitmap,
@@ -208,6 +209,8 @@ export class CommissioningServer extends MatterNode {
 
     private deviceInstance?: MatterDevice;
     private interactionServer?: InteractionServer;
+
+    protected readonly rootEndpoint = new RootEndpoint();
 
     private nextEndpointId: EndpointNumber;
 
@@ -422,13 +425,81 @@ export class CommissioningServer extends MatterNode {
     }
 
     /**
+     * Get a cluster server from the root endpoint. This is mainly used internally and not needed to be called by the user.
+     *
+     * @param cluster ClusterServer to get or undefined if not existing
+     */
+    getRootClusterServer<
+        F extends BitSchema,
+        SF extends TypeFromPartialBitSchema<F>,
+        A extends Attributes,
+        C extends Commands,
+        E extends Events,
+    >(cluster: Cluster<F, SF, A, C, E>): ClusterServerObj<A, E> | undefined {
+        return this.rootEndpoint.getClusterServer(cluster);
+    }
+
+    /**
+     * Add a cluster client to the root endpoint. This is mainly used internally and not needed to be called by the user.
+     *
+     * @param cluster ClusterClient object to add
+     */
+    addRootClusterClient<F extends BitSchema, A extends Attributes, C extends Commands, E extends Events>(
+        cluster: ClusterClientObj<F, A, C, E>,
+    ) {
+        this.rootEndpoint.addClusterClient(cluster);
+    }
+
+    /**
+     * Get a cluster client from the root endpoint. This is mainly used internally and not needed to be called by the user.
+     *
+     * @param cluster ClusterClient to get or undefined if not existing
+     */
+    getRootClusterClient<
+        F extends BitSchema,
+        SF extends TypeFromPartialBitSchema<F>,
+        A extends Attributes,
+        C extends Commands,
+        E extends Events,
+    >(cluster: Cluster<F, SF, A, C, E>): ClusterClientObj<F, A, C, E> | undefined {
+        return this.rootEndpoint.getClusterClient(cluster);
+    }
+
+    /**
+     * Get the root endpoint of the node.
+     */
+    getRootEndpoint() {
+        return this.rootEndpoint;
+    }
+
+    /**
+     * Add a child endpoint to the root endpoint. This is mainly used internally and not needed to be called by the user.
+     *
+     * @param endpoint Endpoint to add
+     * @protected
+     */
+    protected addEndpoint(endpoint: Endpoint) {
+        this.rootEndpoint.addChildEndpoint(endpoint);
+    }
+
+    /**
+     * Get a child endpoint from the root endpoint. This is mainly used internally and not needed to be called by the user.
+     *
+     * @param endpointId Endpoint ID of the child endpoint to get
+     * @protected
+     */
+    protected getChildEndpoint(endpointId: EndpointNumber): Endpoint | undefined {
+        return this.rootEndpoint.getChildEndpoint(endpointId);
+    }
+
+    /**
      * Add a new cluster server to the root endpoint
      * BasicInformationCluster and OperationalCredentialsCluster can not be added via this method because they are
      * added in the constructor
      *
      * @param cluster
      */
-    override addRootClusterServer<A extends Attributes, E extends Events>(cluster: ClusterServerObj<A, E>) {
+    addRootClusterServer<A extends Attributes, E extends Events>(cluster: ClusterServerObj<A, E>) {
         if (cluster.id === BasicInformationCluster.id) {
             throw new ImplementationError(
                 "BasicInformationCluster can not be modified, provide all details in constructor options!",
@@ -439,7 +510,7 @@ export class CommissioningServer extends MatterNode {
                 "OperationalCredentialsCluster can not be modified, provide the certificates in constructor options!",
             );
         }
-        super.addRootClusterServer(cluster);
+        this.rootEndpoint.addClusterServer(cluster);
     }
 
     /**
